RecordActions: split actions into multiple files
